serverless-calendar/event/requirements.txt
slackclient==2.5.0
python-dateutil==2.8.1
firebase_admin==3.1.0
google-cloud-tasks==1.5.0
googleapis-common-protos==1.51.0
